+++ datamule/docs/source/changelog.rst
@@ -1,5 +1,9 @@
 Changelog
 =========
+v.428 (01-03-2025)
+- added option to set download source via Config class
+- added option to search by document type and by containing text to Portfolio class
+- added download option to Portfolio class. Now you don't need to import Downloader class to download submissions
 
 v0.418 (2024-12-17)
 - improved imports